fix kan-bayashi review and move e2e to rnn
