+++ espnet/nets/pytorch_backend/e2e_asr_transformer.py
@@ -248,7 +248,7 @@ class E2E(ASRInterface, torch.nn.Module):
             hyp['ctc_score_prev'] = 0.0
             if ctc_weight != 1.0:
                 # pre-pruning based on attention scores
-                from espnet.nets.pytorch_backend.decoders import CTC_SCORING_RATIO
+                from espnet.nets.pytorch_backend.rnn.decoders import CTC_SCORING_RATIO
                 ctc_beam = min(lpz.shape[-1], int(beam * CTC_SCORING_RATIO))
             else:
                 ctc_beam = lpz.shape[-1]
